+++ pybamm/models/submodels/electrode/ohm/leading_ohm.py
@@ -37,12 +37,14 @@ class LeadingOrder(BaseModel):
         """
         Returns variables which are derived from the fundamental variables in the model.
         """
+        param = self.param
+
         i_boundary_cc = variables["Current collector current density"]
         phi_s_cn = variables["Negative current collector potential"]
 
         # import parameters and spatial variables
-        l_n = self.param.l_n
-        l_p = self.param.l_p
+        l_n = param.l_n
+        l_p = param.l_p
         x_n = pybamm.standard_spatial_vars.x_n
         x_p = pybamm.standard_spatial_vars.x_p