+++ tests/unit/test_citations.py
@@ -123,14 +123,14 @@ class TestCitations(unittest.TestCase):
         citations._reset()
         self.assertNotIn("Subramanian2005", citations._papers_to_cite)
         pybamm.particle.no_distribution.XAveragedPolynomialProfile(
-            None, "Negative", "quadratic profile"
+            None, "Negative", "quadratic profile", None
         )
         self.assertIn("Subramanian2005", citations._papers_to_cite)
 
         citations._reset()
         self.assertNotIn("Subramanian2005", citations._papers_to_cite)
         pybamm.particle.no_distribution.PolynomialProfile(
-            None, "Negative", "quadratic profile"
+            None, "Negative", "quadratic profile", None
         )
         self.assertIn("Subramanian2005", citations._papers_to_cite)
 
@@ -254,7 +254,7 @@ class TestCitations(unittest.TestCase):
             pybamm.IDAKLUSolver()
             self.assertIn("Hindmarsh2005", citations._papers_to_cite)
 
-    @unittest.skipIf(not pybamm.have_jax(), "jax is not installed")
+    @unittest.skipIf(not pybamm.have_jax(), "jax or jaxlib is not installed")
     def test_jax_citations(self):
         citations = pybamm.citations
         citations._reset()