+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_spm.py
@@ -161,9 +161,10 @@ class TestSPM(unittest.TestCase):
         self.assertEqual(new_model.timescale.id, model.timescale.id)
 
         # with custom submodels
-        model = pybamm.lithium_ion.SPM({"thermal": "x-full"}, build=False)
+        options = {"stress-induced diffusion": "false", "thermal": "x-full"}
+        model = pybamm.lithium_ion.SPM(options, build=False)
         particle_n = pybamm.particle.no_distribution.XAveragedPolynomialProfile(
-            model.param, "Negative", "quadratic profile"
+            model.param, "Negative", "quadratic profile", options
         )
         model.submodels["negative particle"] = particle_n
         model.build_model()
@@ -253,6 +254,14 @@ class TestSPMWithMechanics(unittest.TestCase):
 
     def test_well_posed_both_swelling_only(self):
         options = {"particle mechanics": "swelling only"}
+        model = pybamm.lithium_ion.SPM(options)
+        model.check_well_posedness()
+
+    def test_stress_induced_diffusion_mixed(self):
+        options = {
+            "particle mechanics": "swelling only",
+            "stress-induced diffusion": ("true", "false"),
+        }
         model = pybamm.lithium_ion.SPM(options)
         model.check_well_posedness()