+++ tests/unit/test_parameters/test_parameter_values.py
@@ -4,6 +4,7 @@
 
 import os
 import tempfile
+import shutil
 import subprocess
 import unittest
 
@@ -16,6 +17,14 @@ import tests.shared as shared
 
 
 class TestParameterValues(unittest.TestCase):
+    def tearDown(self):
+        # Make sure the local lithium_ion directory is removed
+        try:
+            shutil.rmtree("lithium_ion")
+        except FileNotFoundError:
+            pass
+        return super().tearDown()
+
     def test_find_parameter(self):
         f = tempfile.NamedTemporaryFile()
         pybamm.PARAMETER_PATH.append(tempfile.gettempdir())
@@ -81,8 +90,7 @@ class TestParameterValues(unittest.TestCase):
         pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Chen2020)
 
         # Clean up parameter files
-        cmd = ["rm", "-r", "lithium_ion"]
-        subprocess.run(cmd)
+        shutil.rmtree("lithium_ion")
 
     def test_update(self):
         # converts to dict if not